# specs/openapi/api.v1.yaml
openapi: "3.1.0"
info:
  title: "LongoX Platform API"
  version: "1.0.0"
  description: "LongoX API — low-code automation and internal tools platform"
  contact:
    name: "LongoX Support"
    url: "https://longox.io/support"
  license:
    name: "Proprietary"
servers:
  - url: "https://api.longox.io"
    description: "Production"
  - url: "https://staging-api.longox.io"
    description: "Staging"
  - url: "http://localhost:3000"
    description: "Development"
paths:
...
  /api/v1/executions:
    get:
      operationId: listExecutions
      tags: ["Executions"]
      summary: "List executions"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
        - $ref: "#/components/parameters/ExecutionWorkflowFilter"
        - $ref: "#/components/parameters/ExecutionStatusFilter"
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: "List of executions"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Execution"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
    post:
      operationId: createExecution
      tags: ["Executions"]
      summary: "Start a new execution"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExecutionInput"
      responses:
        "201":
          description: "Execution started"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /api/v1/executions/{id}:
    get:
      operationId: getExecution
      tags: ["Executions"]
      summary: "Get execution details"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Execution details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/executions/{id}/approve:
    post:
      operationId: approveExecutionTask
      tags: ["Executions"]
      summary: "Approve or reject an approval task"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveTaskInput"
      responses:
        "200":
          description: "Approval processed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalTaskResult"

  # ─── Connectors ─────────────────────────────────────────────────────────
...
  /api/v1/templates:
    get:
      operationId: listTemplates
      tags: ["Templates"]
      summary: "List templates"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
        - $ref: "#/components/parameters/TemplateCategoryFilter"
        - $ref: "#/components/parameters/TemplateSearch"
      responses:
        "200":
          description: "Templates"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
                  total:
                    type: integer
    post:
      operationId: createTemplate
      tags: ["Templates"]
      summary: "Create a template"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateInput"
      responses:
        "201":
          description: "Created template"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
...
  /api/v1/search:
    get:
      operationId: search
      tags: ["Search"]
      summary: "Global search"
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
      responses:
        "200":
          description: "Search results"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  # ─── Compliance ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from /auth/login or /auth/register"
  parameters:
    TenantHeader:
      name: x-tenant-id
      in: header
      schema:
        type: string
      description: "Tenant context for multi-tenant requests"
    PaginationPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PaginationLimit:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ExecutionWorkflowFilter:
      name: workflowId
      in: query
      schema:
        type: string
    ExecutionStatusFilter:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, running, success, failed, cancelled, timed_out]
...
    TemplateCategoryFilter:
      name: category
      in: query
      schema:
        type: string
    TemplateSearch:
      name: search
      in: query
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: "Not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
...
    Execution:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        workflowId:
          type: string
        workflowVersionId:
          type: string
        status:
          type: string
        triggerType:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ExecutionDetail:
      allOf:
        - $ref: "#/components/schemas/Execution"
        - type: object
          properties:
            checkpoints:
              type: array
              items:
                $ref: "#/components/schemas/ExecutionCheckpoint"
            logs:
              type: array
              items:
                $ref: "#/components/schemas/ExecutionLog"
    CreateExecutionInput:
      type: object
      required: [workflowId]
      properties:
        workflowId:
          type: string
        workflowVersionId:
          type: string
        input:
          type: object
        environment:
          type: string
    ExecutionLog:
      type: object
      properties:
        id:
          type: string
        executionId:
          type: string
        level:
          type: string
        message:
          type: string
        nodeId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
    ExecutionCheckpoint:
      type: object
      properties:
        id:
          type: string
        executionId:
          type: string
        nodeId:
          type: string
        attempt:
          type: integer
        stateJson:
          type: object
        createdAt:
          type: string
          format: date-time
    ApproveTaskInput:
      type: object
      required: [taskId, decision]
      properties:
        taskId:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        note:
          type: string
    ApprovalTaskResult:
      type: object
      properties:
        success:
          type: boolean
        decision:
          type: string
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        visibility:
          type: string
        status:
          type: string
        sourceType:
          type: string
        installCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTemplateInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        sourceType:
          type: string
          default: workflow
        manifestJson:
          type: object
...
    SearchResults:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              title:
                type: string
              description:
                type: string
                nullable: true
              url:
                type: string
              source:
                type: object
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...